% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_label_counts.R
\name{fct_label_counts}
\alias{fct_label_counts}
\title{Recode a factor level to include sample size}
\usage{
fct_label_counts(f)
}
\arguments{
\item{f}{A factor variable}
}
\value{
A factor variable with the counts included in the levels
}
\description{
Recode a factor level to include sample size
}
\examples{
f <- fct_label_counts(iris$Species)
levels(f)
# "setosa (n=50)" "versicolor (n=50)" "virginica (n=50)"
}
