% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_recode_if.R
\name{fct_recode_if}
\alias{fct_recode_if}
\title{Recode a factor level on a condition}
\usage{
fct_recode_if(f, cond, level)
}
\arguments{
\item{f}{A factor variable}

\item{cond}{A predicate to determine when the level should be recoded}

\item{level}{The level created based on the condition}
}
\value{
A factor variable with the new level added in
}
\description{
Recode a factor level on a condition
}
\examples{
df <- data.frame(x = factor(c("apple", "bear", "banana", "dear")),
                 y = factor(c("fruit", "animal", "fruit", "animal")))

dplyr::mutate(df, x = fct_recode_if(x, y == "animal", "not a fruit"))
}
