% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_to_num.R
\name{fct_to_num}
\alias{fct_to_num}
\title{Convert a factor variable with numeric levels to a numeric variable}
\usage{
fct_to_num(f)
}
\arguments{
\item{f}{A factor variable with numeric levels}
}
\value{
A numeric variable
}
\description{
Convert a factor variable with numeric levels to a numeric variable
}
\examples{
var <- factor(c("1", "2", "3"))
fct_to_num(var)
}
