% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_trim.R
\name{fct_trim}
\alias{fct_trim}
\title{Trim whitespace from factor levels}
\usage{
fct_trim(f, ...)
}
\arguments{
\item{f}{A factor variable with whitespace}

\item{...}{Arguments passed on to `stringr::str_trim()`}
}
\value{
A factor variable with whitespace removed
}
\description{
Trim whitespace from factor levels
}
\examples{
var <- factor(c("  Level with trailing and leading white space\t",
"\n\nLevel with trailing and leading white space\n\n"))
fct_trim(var)
}
