% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wrap.R
\name{fct_wrap}
\alias{fct_wrap}
\title{Wrap factor levels into nicely formatted paragraphs}
\usage{
fct_wrap(f, ...)
}
\arguments{
\item{f}{A factor variable with long levels}

\item{...}{Arguments passed on to `stringr::str_wrap()`}
}
\value{
A factor variable with re-wrapped levels
}
\description{
Wrap factor levels into nicely formatted paragraphs
}
\examples{
var <- factor(c("a very, very, very, very long factor level", "an even longer, unbelievably longer, factor level"))
fct_wrap(var, width = 10)
}
