% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcatsmisc-package.r
\docType{package}
\name{forcatsmisc}
\alias{forcatsmisc}
\title{forcatsmisc: Additional functions for forcats}
\description{
There are some additional ways that the handling of factor levels can be made easier,
but they are not deemed worthy to be part of the forcats package. They will live here
instead.
}
