% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl_number.R
\name{lvl_number}
\alias{lvl_number}
\title{Return the numeric position of a level in a factor variable}
\usage{
lvl_number(f, lvl)
}
\arguments{
\item{f}{A factor variable}

\item{lvl}{The level for which the underlying value is returned}
}
\value{
The numeric position of the level
}
\description{
Return the numeric position of a level in a factor variable
}
\examples{

lvl_number(iris$Species, "versicolor")
}
