% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl_range.R
\name{lvl_range}
\alias{lvl_range}
\title{Return a sub-range of levels}
\usage{
lvl_range(f, lvls)
}
\arguments{
\item{f}{A factor variable}

\item{lvls}{A vector of 2 levels that you want the}
}
\value{
A vector of the factor variable's levels located in between the levels specified, inclusive.
}
\description{
Return a sub-range of levels
}
\examples{
var <- factor(c("Very likely", "Somewhat likely", "Somewhat unlikely", "Very unlikely"))
lvl_range(var, c("Very likely", "Somewhat unlikely"))
}
